Switch creator.
I don't intend to take full credit as Philipp von Weitershausen is the original creator of this project, but I'm the creator of the "Plus" fork, and when you install it it lists his name instead of mine despite the fact that it links to my extension page. I do intend to eventually rewrite all the guts of this project as a lot of it is legacy code that is broken.